Revert "Merge pull request #619 from alphagov/BAU/move-matchers-to-sharedtest-module"
This reverts commit c544c82a8b362790dec985a1fd31ac6e9001ba59, reversing
changes made to 83bd3d9672063a7f8aa6e3c736044e558c6133f3.

This change is breaking the project in IntelliJ; for some reason the shared-test dependencies do not resolve.  This problem doesn't seem to occur when running the pre-commit script.
