+++ account-management-api/build.gradle
@@ -21,8 +21,7 @@ dependencies {
     runtimeOnly configurations.logging_runtime
 
     testImplementation configurations.tests,
-            configurations.lambda_tests,
-            project(":shared-test")
+            configurations.lambda_tests
 
     testRuntimeOnly configurations.test_runtime
 }
