+++ frontend-api/build.gradle
@@ -19,8 +19,7 @@ dependencies {
     runtimeOnly configurations.logging_runtime
 
     testImplementation configurations.tests,
-            configurations.lambda_tests,
-            project(":shared-test")
+            configurations.lambda_tests
 
     testRuntimeOnly configurations.test_runtime
 }
