+++ shared-test/build.gradle
@@ -11,11 +11,8 @@ repositories {
 
 dependencies {
     implementation configurations.tests,
-            configurations.lambda,
             configurations.glassfish,
-            "com.google.protobuf:protobuf-java:3.18.0",
-            "org.eclipse.jetty:jetty-server:11.0.6",
-            project(":shared")
+            "org.eclipse.jetty:jetty-server:11.0.6"
 }
 
 test {
