+++ shared/build.gradle
@@ -24,8 +24,7 @@ dependencies {
             "com.google.protobuf:protobuf-java:3.18.0"
 
     testImplementation configurations.tests,
-            configurations.lambda_tests,
-            project(":shared-test")
+            configurations.lambda_tests
     testRuntimeOnly configurations.test_runtime
 }
