+++ shared/src/main/java/uk/gov/di/authentication/shared/matchers/APIGatewayProxyResponseEventMatcher.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.sharedtest.matchers;
+package uk.gov.di.authentication.shared.matchers;
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
 import com.fasterxml.jackson.core.JsonProcessingException;
