+++ shared/src/main/java/uk/gov/di/authentication/shared/matchers/AuditMessageMatcher.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.sharedtest.matchers;
+package uk.gov.di.authentication.shared.matchers;
 
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeDiagnosingMatcher;
